Adds unix epoch time to finished tasks.
This commit appends the unix epoch time to
the text of a task when it is marked as done.
It also recomputes the hash of the task
based on the new text before adding it
to the file of completed tasks.

This solves the problem that if you are completing
a task that matches the text of an already
completed task, it is not added to the text file.

This is probably not an issue in light of how
the program was originally intended to work
as designed by Steve Losh. But I want to keep
statistics of my completed tasks, so I want
a way for dupes to be added.